final/requirements.txt
boto3==1.26.3
botocore==1.29.3
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
protobuf==3.20.1
pymongo==4.3.2
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.21.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.42
urllib3==1.26.12
Werkzeug==2.2.2
